增加微信支付通知

+++ src/main/scala/com/dounine/douyinpay/router/routers/WechatRouter.scala
@@ -251,13 +251,13 @@ class WechatRouter()(implicit system: ActorSystem[_])
                                                 | - 总支付金额：${wechatPays
                                         .filter(_.pay == PayStatus.payed)
                                         .map(_.money / 100D)
-                                        .sum + (notifyResponse.total_fee / 100D)} 元
+                                        .sum + (notifyResponse.total_fee.toInt / 100D)} 元
                                                 | - 总支付次数：${wechatPays
                                         .filter(_.pay == PayStatus.payed)
                                         .size + 1}
                                                 | - 帐户余额：${vipUser
                                         .map(_.money / 100D)
-                                        .getOrElse(0D) + (notifyResponse.total_fee / 100D)}
+                                        .getOrElse(0D) + (notifyResponse.total_fee.toInt / 100D)}
                                                 | - 注册时间：${java.time.Duration
                                         .between(
                                           userInfo.get.createTime,
